feat: implement entry join logic with sequence locking and result generation

+++ src/main/java/com/jhj/lottoevent/repository/ResultRepository.java
@@ -3,5 +3,5 @@ package com.jhj.lottoevent.repository;
 import com.jhj.lottoevent.domain.result.Result;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResultRepository extends JpaRepository<Result, Integer> {
+public interface ResultRepository extends JpaRepository<Result, Long> {
 }
