+++ src/main/java/com/jhj/lottoevent/repository/ResultViewRepository.java
@@ -3,5 +3,5 @@ package com.jhj.lottoevent.repository;
 import com.jhj.lottoevent.domain.result.ResultView;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResultViewRepository extends JpaRepository<ResultView, Integer> {
+public interface ResultViewRepository extends JpaRepository<ResultView, Long> {
 }
