+++ src/main/java/com/jhj/lottoevent/repository/SmsLogRepository.java
@@ -3,5 +3,5 @@ package com.jhj.lottoevent.repository;
 import com.jhj.lottoevent.domain.sms.SmsLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SmsLogRepository extends JpaRepository<SmsLog, Integer> {
+public interface SmsLogRepository extends JpaRepository<SmsLog, Long> {
 }
